Remove  field as it's used by node and the browser
This is used by webpack for both node and browser targets. As we need different shims/node-fetch version, it's better
to just ship a browser build (esm) and a node build (cjs, shims external).

This may change in the future, but this is just bike-shedding now.

It's almost certainly better to make the user bring a  compatible API client and pass it in, recommending
to set their https.Agent config correctly when they do so, than handling it for them